+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## (next)
+## 0.36.0
 
 * Introduce `AnyColumnView::Text` and `AnyColumnView::WText` now holds a `TextColumnView` instead of a `TextColumnIt`. To get the iterator you can just call `iter()` on the `TextColumnView`.
 * Introduce `AnyColumnView::Binary` now holds a `BinColumnView` instead of a `BinColumnIt`.
@@ -8,6 +8,8 @@
 * Add `AnyColumnView::as_w_text_view`.
 * Add `AnyColumnView::as_slice`.
 * Add `AnyColumnView::as_nullable_slice`.
+
+The `TextColumnView` has been added in order to allow for faster bindings to arrow. As it gives access to the raw value buffer.
 
 ## 0.35.2